Removed DOS line endings.
Added do_method=whoami to pubsub.py

Note that pubsub.py does at present work.
Someone needs to take the time to go through it carefully.

Since you're doing that, check and see if any of the other .py files
included in this directory are also in the standard distribution of
Python, and if so, remove them from here so people can just use the
standard versions.