Renamed from "PHP PubSub Client Library" to "PubSub Client Library for PHP" after reading the PHP license.
Now it works even in newer PHP4, where globals are not imported
(yay!); added a kn__gpc() function to emulate globals in a
backwards-compatible fashion.

Now each page displays the PHP version.

Added a PUBSUB_NULL constant which is NULL in PHP4 and "" in PHP3.

Switched from __call__ to __call for the callable interface to be
compatible with PHP4 object overloading.

Now kn_apply works in both PHP3 and PHP4.

Prefixed the nasty $kn_server "local" globals with underscores.

Added "NULL" support to kn_toSource.